+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 12.5
 - SQL fix #359 duplicate entries when school is renamed after Rollover in AddDrop.php, thanks to @HubertQuiquet
 - SQL case-insensitive username uniqueness check in ImportUsers.fnc.php
 - Fix HTML error label's for attribute doesn't match any element id in General_Info.php
+- Fix add .tinymce CSS class when class already set in $extra param in Inputs.php
 
 Changes in 12.4.2
 -----------------